requirements.txt
requests==2.18.4
gtts==2.2.3
pydub==0.25.1
PyYAML==6.0
python-telegram-bot --upgrade
importlib-resources